# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Compras/DetalleCompra": {
      "get": {
        "tags": [
          "Compras"
        ],
        "operationId": "GetDetalleHerramientasParaCompra",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComprasParaDetalleDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComprasParaDetalleDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComprasParaDetalleDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Compras/CrearCompra": {
      "post": {
        "tags": [
          "Compras"
        ],
        "operationId": "CrearCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCompraDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCompraDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCompraDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComprasParaDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComprasParaDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComprasParaDetalleDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompraItemsDTO": {
        "type": "object",
        "properties": {
          "idHerramienta": {
            "type": "integer",
            "format": "int32"
          },
          "nombreHerramienta": {
            "type": "string",
            "nullable": true
          },
          "materialHerramienta": {
            "type": "string",
            "nullable": true
          },
          "precioHerramienta": {
            "type": "number",
            "format": "float"
          },
          "descripcionHerramienta": {
            "type": "string",
            "nullable": true
          },
          "cantidadHerramienta": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ComprasParaDetalleDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          },
          "direccionEnvio": {
            "type": "string",
            "nullable": true
          },
          "precioTotal": {
            "type": "number",
            "format": "float"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemsDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CrearCompraDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          },
          "metodoPagoId": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemsDTO"
            },
            "nullable": true
          },
          "direccionEnvio": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}